Changed bytes to ints (no real advantage of using byte) and improvements to the PetriNet Class (renamed from NetCreator)

+++ piaf/petrinet/Place.java
@@ -54,7 +54,7 @@ public class Place {
 	 * @param BaseTransisition transition
 	 */
 	public boolean setTransition(Transition transition) {
-		byte i = 0;
+		int i = 0;
 		// Find next empty transition
 		while ((Transitions[i] != null) && (i < Nr_Of_Transitions_Out)) {
 			i += 1;
